sorter_v2_dev_only/requirements.txt
# Sorter 2.0 Requirements
# Core dependencies (required)
pillow>=9.0.0

# GUI dependencies (optional - for enhanced UI)
customtkinter>=5.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
